Fix PriorityQueue.Count type, rename TroopManagment.cs, clear selectedTroops, remove [ReadOnly] from walkCosts, fix TroopCombatSystem.Die() faction handling
Agent-Logs-Url: https://github.com/kwanyinlai/tower-defense-game/sessions/3c7c630f-29ea-4a12-8483-041905d6ad46

+++ tower-defense-game/Assets/Scripts/PriorityQueue.cs
@@ -9,7 +9,7 @@ public class PriorityQueue<T>
 {
     private readonly List<(T Item, float Priority)> _heap = new();
 
-    public float Count => _heap.Count;
+    public int Count => _heap.Count;
 
     public void Enqueue(T item, float priority)
     {
